+++ src/components/Routes/Routes.tsx
@@ -2,7 +2,7 @@ import React, { FC } from "react";
 import { Route, Switch } from "react-router-dom";
 
 import MakePage from "../../pages/Make/Make";
-import MySwapsPage from "../../pages/MyOrders/MySwaps";
+import MySwapsPage from "../../pages/MyOrders/MyOrders";
 import OrderDetail from "../../pages/OrderDetail/OrderDetail";
 import SwapPage from "../../pages/Swap/Swap";
 import { AppRoutes } from "../../routes";
