+++ src/pages/MyOrders/MyOrders.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from "react";
 
 import MyOrdersWidget from "../../components/MyOrdersWidget/MyOrdersWidget";
-import { StyledPage } from "./MySwaps.styles";
+import { StyledPage } from "./MyOrders.styles";
 
 const MySwapsPage: FC = () => {
   return (
